Added TypeScript definitions file
This file is not 100% complete, in that it ignores the instanced usage described in the documentation, but that usage does not appear to work properly. The module only exports an instance of the Log class, it does not export the declaration of that class, and so attempting to create a new instance of it using `new Log(1)` throws `TypeError: Log is not a constructor`.

It also simplifies the `log.Log` property down to its intended usage, which is an enum for the log levels.